Remove backend dependency on Xamarin
