+++ .gitignore
@@ -286,3 +286,4 @@ Resource.designer.cs
 /DotNetRu.Azure/pushconfig.json
 /DotNetRu.AzureService/appsettings.Production.json
 /DotNetRu.AzureService/Properties/ServiceDependencies/DotNetRuAppService - Web Deploy/profile.arm.json
+.DS_Store
