+++ DotNetRu.AppUtils/DotNetRu.AppUtils.csproj
@@ -1,0 +1,29 @@
+<Project Sdk="Microsoft.NET.Sdk">
+
+  <PropertyGroup>
+    <TargetFramework>netstandard2.1</TargetFramework>
+    <Nullable>enable</Nullable>
+    <LangVersion>10.0</LangVersion>
+  </PropertyGroup>
+
+  <ItemGroup>
+    <PackageReference Include="Xamarin.Essentials" Version="1.6.1" />
+    <PackageReference Include="Xamarin.Forms" Version="5.0.0.2012" />
+    <PackageReference Include="Microsoft.AppCenter.Analytics" Version="3.4.3" />
+    <PackageReference Include="Microsoft.AppCenter.Crashes" Version="3.4.3" />
+  </ItemGroup>
+
+  <ItemGroup>
+    <None Update="Config\config.DEV.json">
+      <DependentUpon>..\..\DotNetRu.Utils\Config\config.json</DependentUpon>
+    </None>
+    <None Update="Config\config.PROD.json">
+      <DependentUpon>..\..\DotNetRu.Utils\Config\config.json</DependentUpon>
+    </None>
+  </ItemGroup>
+  <ItemGroup>
+    <EmbeddedResource Include="Config\config.json">
+      <LogicalName>config.json</LogicalName>
+    </EmbeddedResource>
+  </ItemGroup>
+</Project>
