+++ DotNetRu.AppUtils/Logging/DotNetRuLoggerKeys.cs
@@ -1,0 +1,16 @@
+namespace DotNetRu.AppUtils.Logging;
+
+public static class DotNetRuLoggerKeys
+{
+    public const string ReminderAdded = "ReminderAdded";
+
+    public const string ReminderRemoved = "ReminderRemoved";
+
+    public const string Share = "Share";
+
+    public const string LeaveFeedback = "LeaveFeedback";
+
+    public const string ConferenceFeedback = "ConferenceFeedback";
+
+    public const string LaunchedBrowser = "LaunchedBrowser";        
+}
