+++ DotNetRu.AzureService/Helpers/RealmHelper.cs
@@ -1,12 +1,8 @@
 using System;
-using System.Collections.Generic;
 using System.IO;
 using System.Linq;
 using System.Threading.Tasks;
-using DotNetRu.DataStore.Audit.Helpers;
-using DotNetRu.DataStore.Audit.Models;
 using DotNetRu.DataStore.Audit.RealmModels;
-using DotNetRu.Models.Social;
 using Realms;
 using Realms.Sync;
 
@@ -26,7 +22,7 @@ namespace DotNetRu.AzureService.Helpers
         {
             SyncConfigurationBase.Initialize(UserPersistenceMode.NotEncrypted, basePath: Path.GetTempPath());
 
-            return await Realms.Sync.User.LoginAsync(
+            return await User.LoginAsync(
                 Credentials.UsernamePassword(realmSettings.Login, realmSettings.Password, createUser: false),
                 new Uri($"https://{realmSettings.RealmServerUrl}"));
         }
@@ -36,15 +32,6 @@ namespace DotNetRu.AzureService.Helpers
             var auditVersion = realm.All<AuditVersion>();
 
             return auditVersion.Single().CommitHash;
-        }
-
-        internal static async Task<IEnumerable<SubscriptionModel>> GetAllCommunitiesAsync(this RealmSettings realmSettings, SocialMediaType socialMediaType)
-        {
-            var user = await realmSettings.GetUser();
-            var realmUrl = new Uri($"realms://{realmSettings.RealmServerUrl}/{realmSettings.RealmName}");
-            var realm = await user.GetRealm(realmUrl);
-            return SubscriptionsHelper.GetDefaultCommunitySubscriptionsByRealm(realm)
-                .Where(x => x.Type == socialMediaType);
         }
     }
 }
