+++ DotNetRu.Clients.UI/About/Friends/FriendsViewModel.cs
@@ -3,11 +3,10 @@ namespace DotNetRu.Clients.Portable.ViewModel
     using System;
     using System.Linq;
     using System.Windows.Input;
-    using DotNetRu.Clients.UI.Helpers;
+    using DotNetRu.AppUtils;
+    using DotNetRu.AppUtils.Logging;
     using DotNetRu.DataStore.Audit.Models;
     using DotNetRu.DataStore.Audit.Services;
-    using DotNetRu.Utils;
-    using DotNetRu.Utils.Helpers;
 
     using FormsToolkit;
