+++ DotNetRu.Clients.UI/Feedback/FeedbackViewModel.cs
@@ -2,11 +2,9 @@ namespace DotNetRu.Clients.Portable.ViewModel
 {
     using System;
     using System.Windows.Input;
-
+    using DotNetRu.AppUtils;
+    using DotNetRu.AppUtils.Logging;
     using DotNetRu.DataStore.Audit.Models;
-    using DotNetRu.Utils;
-    using DotNetRu.Utils.Helpers;
-    using DotNetRu.Utils.Interfaces;
 
     using FormsToolkit;
