+++ DotNetRu.Clients.UI/Helpers/MessagingUtils.cs
@@ -1,5 +1,5 @@
+using DotNetRu.AppUtils;
 using DotNetRu.Clients.Portable.Interfaces;
-using DotNetRu.Utils.Helpers;
 using FormsToolkit;
 using Xamarin.Forms;
