+++ DotNetRu.Clients.UI/Meetups/FilterSessionsViewModel.cs
@@ -2,9 +2,8 @@ namespace DotNetRu.Clients.Portable.ViewModel
 {
     using System.Collections.Generic;
     using System.Linq;
-
+    using DotNetRu.AppUtils;
     using DotNetRu.DataStore.Audit.Models;
-    using DotNetRu.Utils.Helpers;
     using Xamarin.Forms;
 
     public class FilterSessionsViewModel : ViewModelBase
