+++ DotNetRu.Clients.UI/News/NewsPage.xaml.cs
@@ -4,10 +4,9 @@ using Xamarin.Essentials;
 namespace DotNetRu.Clients.UI.Pages.Home
 {
     using System;
-
+    using DotNetRu.AppUtils;
     using DotNetRu.Clients.Portable.Model;
     using DotNetRu.Clients.Portable.ViewModel;
-    using DotNetRu.Utils.Helpers;
 
     using FormsToolkit;
