+++ DotNetRu.Clients.UI/News/NewsViewModel.cs
@@ -7,11 +7,10 @@ namespace DotNetRu.Clients.Portable.ViewModel
     using System.Linq;
     using System.Threading.Tasks;
     using System.Windows.Input;
-
-    using DotNetRu.Clients.Portable.Model;
+    using DotNetRu.AppUtils;
+    using DotNetRu.AppUtils.Logging;
     using DotNetRu.Clients.Portable.Services;
     using DotNetRu.DataStore.Audit.Models;
-    using DotNetRu.Utils;
     using DotNetRu.Utils.Helpers;
 
     using FormsToolkit;
