+++ DotNetRu.Clients.UI/News/VkontakteService.cs
@@ -1,17 +1,13 @@
 using System.Linq;
 using DotNetRu.Models.Social;
-using Flurl.Http.Content;
-using Newtonsoft.Json;
 
 namespace DotNetRu.Clients.Portable.Services
 {
     using System;
     using System.Collections.Generic;
     using System.Threading.Tasks;
-
-    using DotNetRu.Clients.Portable.Model;
-    using DotNetRu.Clients.UI;
-    using DotNetRu.Utils;
+    using DotNetRu.AppUtils.Config;
+    using DotNetRu.AppUtils.Logging;
     using Flurl.Http;
 
     public static class VkontakteService
