+++ DotNetRu.Clients.UI/Speakers/SpeakerDetailsViewModel.cs
@@ -1,6 +1,7 @@
 using System;
 using System.Collections.Generic;
 using System.Linq;
+using DotNetRu.AppUtils;
 using DotNetRu.DataStore.Audit.Models;
 using DotNetRu.Utils.Helpers;
