+++ DotNetRu.DataStore.Audit/DotNetRu.DataStore.Audit.csproj
@@ -18,8 +18,6 @@
     <PackageReference Include="AutoMapper" Version="10.1.1" />
     <PackageReference Include="morelinq" Version="3.3.2" />
     <PackageReference Include="Realm" Version="5.1.2" />
-    <PackageReference Include="Xamarin.Essentials" Version="1.6.1" />
-    <PackageReference Include="Xamarin.Forms" Version="5.0.0.2012" />
   </ItemGroup>
   <ItemGroup>
     <EmbeddedResource Include="DotNetRuOffline.realm">
@@ -72,5 +70,6 @@
     <ProjectReference Include="..\DotNetRu.Models\DotNetRu.Models.csproj" />
     <ProjectReference Include="..\DotNetRu.Utils\DotNetRu.Utils.csproj" />
     <ProjectReference Include="..\RealmUpdateLibrary\DotNetRu.RealmUpdateLibrary.csproj" />
+    <ProjectReference Include="..\DotNetRu.AppUtils\DotNetRu.AppUtils.csproj" />
   </ItemGroup>
 </Project>
