+++ DotNetRu.Utils/DotNetRu.Utils.csproj
@@ -4,24 +4,10 @@
       <ProduceReferenceAssembly>True</ProduceReferenceAssembly>
   </PropertyGroup>
   <ItemGroup>
-    <EmbeddedResource Include="Config\config.json">
-      <LogicalName>config.json</LogicalName>
-    </EmbeddedResource>
-
-    <None Include="Config\config.DEV.json">
-        <DependentUpon>config.json</DependentUpon>
-    </None>
-    <None Include="Config\config.PROD.json">
-        <DependentUpon>config.json</DependentUpon>
-    </None>
-  </ItemGroup>
-  <ItemGroup>
     <PackageReference Include="Microsoft.AppCenter" Version="3.4.3" />
     <PackageReference Include="Microsoft.AppCenter.Analytics" Version="3.4.3" />
     <PackageReference Include="Microsoft.AppCenter.Crashes" Version="3.4.3" />
     <PackageReference Include="Newtonsoft.Json" Version="12.0.3" />
-    <PackageReference Include="Xamarin.Essentials" Version="1.6.1" />
-    <PackageReference Include="Xamarin.Forms" Version="5.0.0.2012" />
   </ItemGroup>
   <ItemGroup>
     <ProjectReference Include="..\DotNetRu.Models\DotNetRu.Models.csproj" />
