--- DotNetRu.Utils/Helpers/MessageKeys.cs
@@ -1,44 +1,0 @@
-namespace DotNetRu.Utils.Helpers
-{
-    public static class MessageKeys
-    {
-        public const string LanguageChanged = "LanguageChanged";
-
-        public const string FriendsChanged = "FriendsChanged";
-
-        public const string MeetupsChanged = "MeetupsChanged";
-
-        public const string SpeakersChanged = "SpeakersChanged";
-
-        public const string UpdateTitles = "UpdateTitles";
-
-        public const string NavigateToEvent = "navigate_event";
-
-        public const string NavigateToSession = "navigate_session";
-
-        public const string NavigateToSpeaker = "navigate_speaker";
-
-        public const string NavigateToSponsor = "navigate_sponsor";
-
-        public const string NavigateToImage = "navigate_image";
-
-        public const string NavigateLogin = "navigate_login";
-
-        public const string NavigateToConferenceFeedback = "navigate_conferencefeedback";
-
-        public const string Error = "error";
-
-        public const string Connection = "connection";
-
-        public const string LoggedIn = "loggedin";
-
-        public const string Message = "message";
-
-        public const string Question = "question";
-
-        public const string Choice = "choice";
-
-        public const string RealmUpdated = nameof(RealmUpdated);
-    }
-}
-
