+++ DotNetRuBackend.sln
@@ -5,8 +5,6 @@ VisualStudioVersion = 17.6.33417.168
 MinimumVisualStudioVersion = 10.0.40219.1
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "DotNetRu.Utils", "DotNetRu.Utils\DotNetRu.Utils.csproj", "{DFA0F0F6-C396-412E-86A9-E9095F71CC36}"
 EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "DotNetRu.DataStore.Audit", "DotNetRu.DataStore.Audit\DotNetRu.DataStore.Audit.csproj", "{7B4C73DF-8A4A-4AD1-92D9-A064065CDD71}"
-EndProject
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Solution Items", "Solution Items", "{B20890AA-A26B-4C29-B848-D24FFEE1DE7C}"
 	ProjectSection(SolutionItems) = preProject
 		.editorconfig = .editorconfig
@@ -20,8 +18,6 @@ EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "DotNetRu.RealmUpdateLibrary", "RealmUpdateLibrary\DotNetRu.RealmUpdateLibrary.csproj", "{8B68192C-72D3-4F84-93CB-DF0C1984B536}"
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "DotNetRu.AzureService", "DotNetRu.AzureService\DotNetRu.AzureService.csproj", "{F16ADECA-FC75-4779-99A6-D68B18F1D684}"
-EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "DotNetRu.RealmTests", "DotNetRu.RealmTests\DotNetRu.RealmTests.csproj", "{CEA1DA8E-EA4D-488D-824E-0BAB63B5B75B}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -63,30 +59,6 @@ Global
 		{DFA0F0F6-C396-412E-86A9-E9095F71CC36}.Release|x64.Build.0 = Release|Any CPU
 		{DFA0F0F6-C396-412E-86A9-E9095F71CC36}.Release|x86.ActiveCfg = Release|Any CPU
 		{DFA0F0F6-C396-412E-86A9-E9095F71CC36}.Release|x86.Build.0 = Release|Any CPU
-		{7B4C73DF-8A4A-4AD1-92D9-A064065CDD71}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{7B4C73DF-8A4A-4AD1-92D9-A064065CDD71}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{7B4C73DF-8A4A-4AD1-92D9-A064065CDD71}.Debug|ARM.ActiveCfg = Debug|Any CPU
-		{7B4C73DF-8A4A-4AD1-92D9-A064065CDD71}.Debug|ARM.Build.0 = Debug|Any CPU
-		{7B4C73DF-8A4A-4AD1-92D9-A064065CDD71}.Debug|iPhone.ActiveCfg = Debug|Any CPU
-		{7B4C73DF-8A4A-4AD1-92D9-A064065CDD71}.Debug|iPhone.Build.0 = Debug|Any CPU
-		{7B4C73DF-8A4A-4AD1-92D9-A064065CDD71}.Debug|iPhoneSimulator.ActiveCfg = Debug|Any CPU
-		{7B4C73DF-8A4A-4AD1-92D9-A064065CDD71}.Debug|iPhoneSimulator.Build.0 = Debug|Any CPU
-		{7B4C73DF-8A4A-4AD1-92D9-A064065CDD71}.Debug|x64.ActiveCfg = Debug|Any CPU
-		{7B4C73DF-8A4A-4AD1-92D9-A064065CDD71}.Debug|x64.Build.0 = Debug|Any CPU
-		{7B4C73DF-8A4A-4AD1-92D9-A064065CDD71}.Debug|x86.ActiveCfg = Debug|Any CPU
-		{7B4C73DF-8A4A-4AD1-92D9-A064065CDD71}.Debug|x86.Build.0 = Debug|Any CPU
-		{7B4C73DF-8A4A-4AD1-92D9-A064065CDD71}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{7B4C73DF-8A4A-4AD1-92D9-A064065CDD71}.Release|Any CPU.Build.0 = Release|Any CPU
-		{7B4C73DF-8A4A-4AD1-92D9-A064065CDD71}.Release|ARM.ActiveCfg = Release|Any CPU
-		{7B4C73DF-8A4A-4AD1-92D9-A064065CDD71}.Release|ARM.Build.0 = Release|Any CPU
-		{7B4C73DF-8A4A-4AD1-92D9-A064065CDD71}.Release|iPhone.ActiveCfg = Release|Any CPU
-		{7B4C73DF-8A4A-4AD1-92D9-A064065CDD71}.Release|iPhone.Build.0 = Release|Any CPU
-		{7B4C73DF-8A4A-4AD1-92D9-A064065CDD71}.Release|iPhoneSimulator.ActiveCfg = Release|Any CPU
-		{7B4C73DF-8A4A-4AD1-92D9-A064065CDD71}.Release|iPhoneSimulator.Build.0 = Release|Any CPU
-		{7B4C73DF-8A4A-4AD1-92D9-A064065CDD71}.Release|x64.ActiveCfg = Release|Any CPU
-		{7B4C73DF-8A4A-4AD1-92D9-A064065CDD71}.Release|x64.Build.0 = Release|Any CPU
-		{7B4C73DF-8A4A-4AD1-92D9-A064065CDD71}.Release|x86.ActiveCfg = Release|Any CPU
-		{7B4C73DF-8A4A-4AD1-92D9-A064065CDD71}.Release|x86.Build.0 = Release|Any CPU
 		{32A902EE-5F33-4385-A27B-E53FFDAF9A2C}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{32A902EE-5F33-4385-A27B-E53FFDAF9A2C}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{32A902EE-5F33-4385-A27B-E53FFDAF9A2C}.Debug|ARM.ActiveCfg = Debug|Any CPU
@@ -159,30 +131,6 @@ Global
 		{F16ADECA-FC75-4779-99A6-D68B18F1D684}.Release|x64.Build.0 = Release|Any CPU
 		{F16ADECA-FC75-4779-99A6-D68B18F1D684}.Release|x86.ActiveCfg = Release|Any CPU
 		{F16ADECA-FC75-4779-99A6-D68B18F1D684}.Release|x86.Build.0 = Release|Any CPU
-		{CEA1DA8E-EA4D-488D-824E-0BAB63B5B75B}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{CEA1DA8E-EA4D-488D-824E-0BAB63B5B75B}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{CEA1DA8E-EA4D-488D-824E-0BAB63B5B75B}.Debug|ARM.ActiveCfg = Debug|Any CPU
-		{CEA1DA8E-EA4D-488D-824E-0BAB63B5B75B}.Debug|ARM.Build.0 = Debug|Any CPU
-		{CEA1DA8E-EA4D-488D-824E-0BAB63B5B75B}.Debug|iPhone.ActiveCfg = Debug|Any CPU
-		{CEA1DA8E-EA4D-488D-824E-0BAB63B5B75B}.Debug|iPhone.Build.0 = Debug|Any CPU
-		{CEA1DA8E-EA4D-488D-824E-0BAB63B5B75B}.Debug|iPhoneSimulator.ActiveCfg = Debug|Any CPU
-		{CEA1DA8E-EA4D-488D-824E-0BAB63B5B75B}.Debug|iPhoneSimulator.Build.0 = Debug|Any CPU
-		{CEA1DA8E-EA4D-488D-824E-0BAB63B5B75B}.Debug|x64.ActiveCfg = Debug|Any CPU
-		{CEA1DA8E-EA4D-488D-824E-0BAB63B5B75B}.Debug|x64.Build.0 = Debug|Any CPU
-		{CEA1DA8E-EA4D-488D-824E-0BAB63B5B75B}.Debug|x86.ActiveCfg = Debug|Any CPU
-		{CEA1DA8E-EA4D-488D-824E-0BAB63B5B75B}.Debug|x86.Build.0 = Debug|Any CPU
-		{CEA1DA8E-EA4D-488D-824E-0BAB63B5B75B}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{CEA1DA8E-EA4D-488D-824E-0BAB63B5B75B}.Release|Any CPU.Build.0 = Release|Any CPU
-		{CEA1DA8E-EA4D-488D-824E-0BAB63B5B75B}.Release|ARM.ActiveCfg = Release|Any CPU
-		{CEA1DA8E-EA4D-488D-824E-0BAB63B5B75B}.Release|ARM.Build.0 = Release|Any CPU
-		{CEA1DA8E-EA4D-488D-824E-0BAB63B5B75B}.Release|iPhone.ActiveCfg = Release|Any CPU
-		{CEA1DA8E-EA4D-488D-824E-0BAB63B5B75B}.Release|iPhone.Build.0 = Release|Any CPU
-		{CEA1DA8E-EA4D-488D-824E-0BAB63B5B75B}.Release|iPhoneSimulator.ActiveCfg = Release|Any CPU
-		{CEA1DA8E-EA4D-488D-824E-0BAB63B5B75B}.Release|iPhoneSimulator.Build.0 = Release|Any CPU
-		{CEA1DA8E-EA4D-488D-824E-0BAB63B5B75B}.Release|x64.ActiveCfg = Release|Any CPU
-		{CEA1DA8E-EA4D-488D-824E-0BAB63B5B75B}.Release|x64.Build.0 = Release|Any CPU
-		{CEA1DA8E-EA4D-488D-824E-0BAB63B5B75B}.Release|x86.ActiveCfg = Release|Any CPU
-		{CEA1DA8E-EA4D-488D-824E-0BAB63B5B75B}.Release|x86.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
